Deleted (int)

+++ a1/life2D/life2D.c
@@ -10,9 +10,9 @@ int main(int argc, char **argv) {
         fprintf(stderr, "Usage: life2D rows cols states\n");
         return 1;
     } else {
-    	int height = (int)strtol(argv[1], NULL, 10);
-    	int width = (int)strtol(argv[2], NULL, 10);
-    	int states = (int)strtol(argv[3], NULL, 10);
+    	int height = strtol(argv[1], NULL, 10);
+    	int width = strtol(argv[2], NULL, 10);
+    	int states = strtol(argv[3], NULL, 10);
     	int board[height * width];
     	int count = 0;
     	while(scanf("%d", &board[count]) != EOF) {
